Added Test File to test this out in HTML
I may not do it since it may take a great amount of time to look into getting the data over into a web page. Because I have to do more coding and testing etc. But I might do this after the intial data itself is released.  This will be pushed out until after the project is released.