Optimize code
Had the data reports be stored in the correct place for seasonal data.  Moved Test2 to be deleted later on. Moved code out of Test file to the correct file. Moved the test file for TotalNumberOfRidersForTheYearTest to the trash folder. I will get rid of it later on.  The Excel reports can stay there for now. I will get rid of them later on.