Got rid of a bunch of files
Some of them may not be useful and they have issues. I did fix the current file I am working. It was making duplicates. Now, I just need to optimize the ram usage on it. It spikes to high. So I need to set it to use less than that.