Added Some example files
Which I may Delete later. Finally got the code to work again. Just had to test it and make some changes to it. It was not parsing the data correctly. I am not going to put into functions(Or maybe I will later). But this has to be done today and released by Friday. I also need to update my portfilo website as well.